feat(di): Add DependencyInjection.SqLite and DependencyInjection.SqlServer packages
Create two new library projects that provide AddDbContextWithRepositories
in a shared namespace. Consumers switch databases by changing the package
reference — no code changes needed.

New packages:
- Ploch.Data.GenericRepository.EFCore.DependencyInjection.SqLite
  registers SqLiteDbContextCreationLifecycle + DbContext + repositories
- Ploch.Data.GenericRepository.EFCore.DependencyInjection.SqlServer
  registers DefaultDbContextCreationLifecycle + DbContext + repositories

Also fixes DataIntegrationTest to register SqLiteDbContextCreationLifecycle
by default (matching the in-memory SQLite test database), and updates the
SampleApp ConsoleApp to use the library's DI package directly.

Refs: #72
