+++ samples/SampleApp/src/ConsoleApp/Ploch.Data.SampleApp.ConsoleApp.csproj
@@ -6,12 +6,11 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" />
     <PackageReference Include="Microsoft.Extensions.Hosting" />
+    <PackageReference Include="Ploch.Data.GenericRepository.EFCore.DependencyInjection.SqLite" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Data.SQLite\Ploch.Data.SampleApp.Data.SQLite.csproj" />
+    <ProjectReference Include="..\Data\Ploch.Data.SampleApp.Data.csproj" />
   </ItemGroup>
   <ItemGroup>
     <None Update="appsettings.json">
