+++ src/Data.GenericRepository/Data.GenericRepository.EFCore.DependencyInjection.SqlServer/Ploch.Data.GenericRepository.EFCore.DependencyInjection.SqlServer.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>$(TargetFrameworkVersions)</TargetFrameworks>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <Platforms>AnyCPU;x64</Platforms>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\Data.EFCore.SqlServer\Ploch.Data.EFCore.SqlServer.csproj" />
+    <ProjectReference Include="..\Data.GenericRepository.EFCore\Ploch.Data.GenericRepository.EFCore.csproj" />
+  </ItemGroup>
+
+</Project>
